analysis/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.1
python-dotenv==1.0.1
tabulate==0.9.0
